AMB-1733 add cis2 authenticator

+++ e2e/utils/base_test.py
@@ -3,9 +3,10 @@ import uuid
 from typing import List
 
 from lib.apigee import ApigeeService, ApigeeApp, ApigeeProduct
-from lib.authentication import AppRestrictedAuthentication
+from lib.authentication import AppRestrictedAuthentication, Cis2Authentication, LoginUser
 from lib.env import get_auth_url, get_proxy_name, get_service_base_path
-from utils.factories import make_apigee_service, make_app_restricted_app, make_apigee_product
+from utils.constants import cis2_user
+from utils.factories import make_apigee_service, make_app_restricted_app, make_cis2_app, make_apigee_product
 from utils.immunisation_api import ImmunisationApi, parse_location
 from utils.resource import create_an_imms_obj
 
@@ -24,51 +25,60 @@ class ImmunizationBaseTest(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
-        cls.apigee_service = make_apigee_service()
-        display_name = f"test-{get_proxy_name()}"
-
-        product_data = ApigeeProduct(name=str(uuid.uuid4()),
-                                     displayName=display_name,
-                                     # we only use one single product for all auth types
-                                     # TODO(Cis2_AMB-1733) add scopes for Cis2
-                                     # TODO(NhsLogin_AMB-1923) add scopes for NhsLogin
-                                     scopes=["urn:nhsd:apim:app:level3:immunisation-fhir-api"])
-        cls.product = make_apigee_product(cls.apigee_service, product_data)
-        cls.apigee_service.add_proxy_to_product(product_name=cls.product.name, proxy_name=get_proxy_name())
-
         cls.apps = []
         cls.imms_apis = []
-
-        def make_app_data() -> ApigeeApp:
-            _app = ApigeeApp(name=str(uuid.uuid4()))
-            _app.set_display_name(display_name)
-            _app.add_product(cls.product.name)
-            return _app
-
-        # ApplicationRestricted
-        app_data = make_app_data()
-        app_restricted_app, app_res_cfg = make_app_restricted_app(cls.apigee_service, app_data)
-        cls.apps.append(app_restricted_app)
-
-        app_res_auth = AppRestrictedAuthentication(get_auth_url(), app_res_cfg)
+        cls.apigee_service = make_apigee_service()
         base_url = get_service_base_path()
+        try:
+            display_name = f"test-{get_proxy_name()}"
 
-        cls.default_imms_api = ImmunisationApi(base_url, app_res_auth)
-        cls.imms_apis.append(cls.default_imms_api)
+            product_data = ApigeeProduct(name=str(uuid.uuid4()),
+                                         displayName=display_name,
+                                         # we only use one single product for all auth types
+                                         # TODO(Cis2_AMB-1733) add scopes for Cis2
+                                         # TODO(NhsLogin_AMB-1923) add scopes for NhsLogin
+                                         scopes=["urn:nhsd:apim:app:level3:immunisation-fhir-api"])
+            cls.product = make_apigee_product(cls.apigee_service, product_data)
+            cls.apigee_service.add_proxy_to_product(product_name=cls.product.name, proxy_name=get_proxy_name())
 
-        # Cis2
-        # TODO(Cis2_AMB-1733) create an app for Cis2 and append it to the cls.apps,
-        #  then create ImmunisationApi and append it to cls.imms_apis
+            def make_app_data() -> ApigeeApp:
+                _app = ApigeeApp(name=str(uuid.uuid4()))
+                _app.set_display_name(display_name)
+                _app.add_product(cls.product.name)
+                return _app
 
-        # NhsLogin
-        # TODO(NhsLogin_AMB-1923) create an app for NhsLogin and append it to the cls.apps,
-        #  then create ImmunisationApi and append it to cls.imms_apis
+            # ApplicationRestricted
+            app_data = make_app_data()
+            app_res_app, app_res_cfg = make_app_restricted_app(cls.apigee_service, app_data)
+            cls.apps.append(app_res_app)
+
+            app_res_auth = AppRestrictedAuthentication(get_auth_url(), app_res_cfg)
+
+            cls.default_imms_api = ImmunisationApi(base_url, app_res_auth)
+            cls.imms_apis.append(cls.default_imms_api)
+
+            # Cis2
+            app_data = make_app_data()
+            cis2_app, app_res_cfg = make_cis2_app(cls.apigee_service, app_data)
+            cls.apps.append(cis2_app)
+
+            cis2_auth = Cis2Authentication(get_auth_url(), app_res_cfg, LoginUser(username=cis2_user))
+            cis2_imms_api = ImmunisationApi(base_url, cis2_auth)
+            cls.imms_apis.append(cis2_imms_api)
+
+            # NhsLogin
+            # TODO(NhsLogin_AMB-1923) create an app for NhsLogin and append it to the cls.apps,
+            #  then create ImmunisationApi and append it to cls.imms_apis
+        except Exception as e:
+            cls.tearDownClass()
+            raise e
 
     @classmethod
     def tearDownClass(cls):
         for app in cls.apps:
             cls.apigee_service.delete_application(app.name)
-        cls.apigee_service.delete_product(cls.product.name)
+        if hasattr(cls, "product") and cls:
+            cls.apigee_service.delete_product(cls.product.name)
 
     @staticmethod
     def create_immunization_resource(imms_api: ImmunisationApi, resource: dict = None) -> str:
